Check if magisk version is 23000 or less before applying policies
TODO: Add the logic for >= magisk v24

+++ custom/ATVServices.sh
@@ -288,19 +288,25 @@ fi
 
 # Give all mitm services root permissions
 
-for package in $MITMPKG com.android.shell; do
-	packageUID=$(dumpsys package "$package" | grep userId | head -n1 | cut -d= -f2)
-	policy=$(sqlite3 /data/adb/magisk.db "select policy from policies where package_name='$package'")
-	if [ "$policy" != 2 ]; then
-		log -p i -t eMagiskATVService "$package current policy is $policy. Adding root permissions..."
-		if ! sqlite3 /data/adb/magisk.db "DELETE from policies WHERE package_name='$package'" ||
-			! sqlite3 /data/adb/magisk.db "INSERT INTO policies (uid,package_name,policy,until,logging,notification) VALUES($packageUID,'$package',2,0,1,1)"; then
-			log -p i -t eMagiskATVService "ERROR: Could not add $package (UID: $packageUID) to Magisk's DB."
-		fi
-	else
-		log -p i -t eMagiskATVService "Root permissions for $package are OK!"
-	fi
-done
+# Check if magisk version is 23000 or less
+
+if [ "$(magisk -V)" -le 23000 ]; then
+    for package in "$MITMPKG" com.android.shell; do
+        packageUID=$(dumpsys package "$package" | grep userId | head -n1 | cut -d= -f2)
+        policy=$(sqlite3 /data/adb/magisk.db "select policy from policies where package_name='$package'")
+        if [ "$policy" != 2 ]; then
+            log -p i -t eMagiskATVService "$package current policy is $policy. Adding root permissions..."
+            if ! sqlite3 /data/adb/magisk.db "DELETE from policies WHERE package_name='$package'" ||
+                ! sqlite3 /data/adb/magisk.db "INSERT INTO policies (uid,package_name,policy,until,logging,notification) VALUES($packageUID,'$package',2,0,1,1)"; then
+                log -p i -t eMagiskATVService "ERROR: Could not add $package (UID: $packageUID) to Magisk's DB."
+            fi
+        else
+            log -p i -t eMagiskATVService "Root permissions for $package are OK!"
+        fi
+    done
+else
+    echo "Magisk version is higher than 23000. Not checking for magisk's policies."
+fi
 
 # Set mitm mock location permission as ignore
